resolved code review issues
