+++ plugins/action/semodule.py
@@ -7,6 +7,7 @@ __metaclass__ = type
 from ansible.plugins.action import ActionBase
 from ansible.errors import AnsibleError
 from ansible.module_utils._text import to_text
+from ansible.module_utils.six import string_types
 import ansible_collections.community.general.plugins.module_utils.semodule as semodule
 from ansible.utils.hashing import checksum
 import traceback
@@ -46,50 +47,50 @@ class ActionModule(ActionBase):
 
     def run(self, tmp=None, task_vars=None):
         result = super(ActionModule, self).run(tmp, task_vars)
-        self._make_tmp_path()
-        new_module = self._task.copy()
-        new_module_args = self._task.args.copy()
-        state = self._task.args.get('state', 'present')
-        if state not in ['present', 'absent']:
-            raise AnsibleError('state can only be present or absent')
-        file_name = self._task.args.get('src', None)
-        if not file_name:
-            raise AnsibleError('src parameter cannot be null')
-        force = self._task.args.get('force', None)
-        tmp_src = self._connection._shell.join_path(self._connection._shell.tmpdir, 'source')
         try:
+            self._make_tmp_path()
+            new_module = self._task.copy()
+            new_module_args = self._task.args.copy()
+            state = self._task.args.get('state', 'present')
+            if state not in ['present', 'absent']:
+                raise AnsibleError('state can only be present or absent')
+            file_name = self._task.args.get('src', None)
+            if not isinstance(file_name, string_types):
+                raise AnsibleError('src parameter cannot be null')
+            force = self._task.args.get('force', None)
+            tmp_src = self._connection._shell.join_path(self._connection._shell.tmpdir, 'source')
             te_file = self._find_needle('files', file_name)
+            try:
+                policy_def = semodule.parse_module_info(semodule.read_te_file(te_file))
+            except IndexError:
+                raise AnsibleError('Module is missing module definition line')
+            semodule_info = self._get_semodule_info()
+            module_installed = self._module_exist(policy_def['name'], semodule_info)
+            if state == 'present':
+                if not module_installed or force:
+                    check_res = self._check_mode(True, policy_def, result)
+                    if check_res:
+                        return check_res
+                    self._copy_te_file(te_file, tmp_src, task_vars)
+                    new_module_args.update(dict(src=tmp_src))
+                    module_return = self._execute_module(module_name='semodule', module_args=new_module_args, task_vars=task_vars)
+                    result.update(module_return)
+            # no need to copy extra files over to target
+            # when we can remove it from here
+            elif state == 'absent':
+                if module_installed:
+                    check_res = self._check_mode(True, policy_def['name'], result)
+                    if check_res:
+                        return check_res
+                    self._remove_module(policy_def['name'])
+                    result['changed'] = True
+                else:
+                    result['changed'] = False
         except AnsibleError as error:
-            result['failed'] = True
-            result['msg'] = to_text(error)
-            result['exception'] = traceback.format_exc()
+                result['failed'] = True
+                result['msg'] = to_text(error)
+                result['exception'] = traceback.format_exc()
+        finally:
+            # clean up temp path
+            self._remove_tmp_path(self._connection._shell.tmpdir)
             return result
-        try:
-            policy_def = semodule.parse_module_info(semodule.read_te_file(te_file))
-        except IndexError:
-            raise AnsibleError('Module is missing module definition line')
-        semodule_info = self._get_semodule_info()
-        module_installed = self._module_exist(policy_def['name'], semodule_info)
-        if state == 'present':
-            if not module_installed or force:
-                check_res = self._check_mode(True, policy_def, result)
-                if check_res:
-                    return check_res
-                self._copy_te_file(te_file, tmp_src, task_vars)
-                new_module_args.update(dict(src=tmp_src))
-                module_return = self._execute_module(module_name='semodule', module_args=new_module_args, task_vars=task_vars)
-                result.update(module_return)
-        # no need to copy extra files over to target
-        # when we can remove it from here
-        elif state == 'absent':
-            if module_installed:
-                check_res = self._check_mode(True, policy_def['name'], result)
-                if check_res:
-                    return check_res
-                self._remove_module(policy_def['name'])
-                result['changed'] = True
-            else:
-                result['changed'] = False
-        # clean up temp path
-        self._remove_tmp_path(self._connection._shell.tmpdir)
-        return result
